SplitName() returns LONG now instead of WORD to allow really long strings.
git-svn-id: https://svn.aros.org/svn/aros/trunk@1634 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/include/clib/dos_protos.h
@@ -705,7 +705,7 @@ AROS_LP4(LONG, SetVBuf,
     AROS_LPA(LONG  , size, D4),
     struct DosLibrary *, DOSBase, 61, Dos)
 
-AROS_LP5(WORD, SplitName,
+AROS_LP5(LONG, SplitName,
     AROS_LPA(STRPTR, name, D1),
     AROS_LPA(ULONG , seperator, D2),
     AROS_LPA(STRPTR, buf, D3),
